% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did_multiplegt_dyn_all_pl.R
\name{did_multiplegt_dyn_all_pl}
\alias{did_multiplegt_dyn_all_pl}
\title{Core function of DIDmultiplegtDYNallPL}
\usage{
did_multiplegt_dyn_all_pl(
  df,
  outcome,
  group,
  time,
  treatment,
  effects = 1,
  placebo = 0,
  switchers = "",
  only_never_switchers = FALSE
)
}
\arguments{
\item{df}{(dataframe) the estimation dataset. It has to be a \if{html}{\out{<ins>}}balanced panel\if{html}{\out{<ins>}} with a least \if{html}{\out{<ins>}}one never-switcher\if{html}{\out{<ins>}} per baseline treatment.}

\item{outcome}{(char) is the outcome variable.}

\item{group}{(char) is the group variable.}

\item{time}{(char) is the time period variable.}

\item{treatment}{(char) is the treatment variable.}

\item{effects}{(int) gives the number of event-study effects to be estimated.}

\item{placebo}{(int) gives the number of placebo estimators to be computed.}

\item{switchers}{(char in ("", "in", "out")) one may be interested in estimating separately the treatment effect of switchers-in, whose average treatment after they switch is larger than their baseline treatment, and of switchers-out, whose average treatment after they switch is lower than their baseline treatment. In that case, one should run the command first with the \code{switchers = "in"} option, and then with the \code{switchers = "out"} option.}

\item{only_never_switchers}{(logical) if this option is specified, the command estimates the event-study effects using only never-switchers as control units.}
}
\value{
The same output as did_multiplegt_dyn.
}
\description{
Auxiliary function of DIDmultiplegtDYN to retrieve more placebos.
}
\section{Overview}{

The following \href{https://github.com/chaisemartinPackages/did_multiplegt_dyn/blob/main/vignettes/vignette_3.md}{vignette} provides an overview of the motivation and the usage of this function.
}

\examples{
GG <- 10; TT <- 4;
set.seed(0)
df <- as.data.frame(matrix(nrow =GG * TT, ncol = 0))
df$G <- floor((1:nrow(df)-1)/TT) + 1
df$T <- ((1:nrow(df)-1) \%\% TT) + 1
df <- df[order(df$G, df$T), ]
df$D <- ifelse(df$T == TT & runif(n = nrow(df)) > 0.5, 1, 0)
df$Y <- runif(n = nrow(df)) * (1 + df$D)
did <- did_multiplegt_dyn_all_pl(df, "Y", "G", "T", "D", effects = 1, placebo = 2)
}
