% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favara_imbs-data.R
\docType{data}
\name{favara_imbs}
\alias{favara_imbs}
\title{Favara and Imbs (2015)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1157 rows and 7 columns.
}
\usage{
data(favara_imbs)
}
\description{
Favara and Imbs (2015) use 1994-to-2005
county-level data to estimate the effect
of the number of regulations lifted on the
growth of mortgages originated by banks,
and on the growth of houses prices. Their
findings are revisited in de Chaisemartin and D'Haultfoeuille (2024a).
This dataset includes only the first 10 states.
The full dataset is available on GitHub.
}
\references{
Favara and Imbs (2015)
(\href{https://www.aeaweb.org/articles?id=10.1257/aer.20121416}{AER})
}
\keyword{datasets}
