% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.did_multiplegt_dyn_all_pl.R
\name{print.did_multiplegt_dyn_all_pl}
\alias{print.did_multiplegt_dyn_all_pl}
\title{Custom print method for did_multiplegt_dyn_all_pl objects}
\usage{
\method{print}{did_multiplegt_dyn_all_pl}(x, ...)
}
\arguments{
\item{x}{A did_multiplegt_dyn_all_pl object}

\item{...}{Undocumented}
}
\value{
No return, custom print method for did_multiplegt_dyn_all_pl objects. Estimation tables are fetched from the object and displayed in the same style as the Stata did_multiplegt_dyn_all_pl command.
}
\description{
A customized printed display for did_multiplegt_dyn_all_pl output
}
