% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.did_multiplegt_dyn}
\alias{summary.did_multiplegt_dyn}
\title{A summary method for did_multiplegt_dyn}
\usage{
\method{summary}{did_multiplegt_dyn}(object, ...)
}
\arguments{
\item{object}{A did_multiplegt_dyn object}

\item{...}{Undocumented}
}
\value{
No return, custom summary method for did_multiplegt_dyn objects. Estimation tables are fetched from the object and displayed in the same style as the Stata did_multiplegt_dyn command.
}
\description{
A customized summary display for did_multiplegt_dyn output
}
